backend/requirements.txt
django
djangorestframework
osmnx 
pandas
scikit-learn
psycopg2-binary

python-dotenv
django-cors-headers
python-decouple

google-auth
djangorestframework-simplejwt
